Refactor messaging terminology and add MessageReceiver
Changes were made to improve code readability and consistency, notably with respect to variable and class names. The main modification occurred in communication classes where the term 'envelope' was replaced with 'message' to better reflect the content being transferred.

In addition, the DiscoveryBridge class was simplified by removing an unnecessary parenthesis, and the MessageReceiver class was added into the receiver package for future use.

Also, several inspection tools were updated in the 'Project_Default.xml' file to include 'previewFile' to enhance code analysis and verification.

The adjustments seek to enhance the code maintainability and clarity, making it easier to understand and continue ongoing development.